Revise ZookeeperCluster
