+++ src/controller_examples/zookeeper_controller/exec/zookeepercluster.rs
@@ -1,0 +1,118 @@
+// Copyright 2022 VMware, Inc.
+// SPDX-License-Identifier: MIT
+use crate::kubernetes_api_objects::{
+    api_resource::*, common::*, dynamic::*, error::ParseDynamicObjectError, marshal::*,
+    object_meta::*, resource::*,
+};
+use crate::pervasive_ext::string_view::*;
+use crate::zookeeper_controller::spec::zookeepercluster::*;
+use vstd::prelude::*;
+
+verus! {
+
+#[verifier(external_body)]
+pub struct ZookeeperCluster {
+    inner: deps_hack::ZookeeperCluster
+}
+
+impl ZookeeperCluster {
+    pub spec fn view(&self) -> ZookeeperClusterView;
+
+    #[verifier(external_body)]
+    pub fn metadata(&self) -> (metadata: ObjectMeta)
+        ensures
+            metadata@ == self@.metadata,
+    {
+        ObjectMeta::from_kube(self.inner.metadata.clone())
+    }
+
+    #[verifier(external_body)]
+    pub fn spec(&self) -> (spec: ZookeeperClusterSpec)
+        ensures
+            spec@ == self@.spec,
+    {
+        ZookeeperClusterSpec::from_kube(self.inner.spec.clone())
+    }
+
+    #[verifier(external_body)]
+    pub fn api_resource() -> (res: ApiResource)
+        ensures
+            res@.kind == ZookeeperClusterView::kind(),
+    {
+        ApiResource::from_kube(deps_hack::kube::api::ApiResource::erase::<deps_hack::ZookeeperCluster>(&()))
+    }
+
+    // NOTE: This function assumes serde_json::to_string won't fail!
+    #[verifier(external_body)]
+    pub fn to_dynamic_object(self) -> (obj: DynamicObject)
+        ensures
+            obj@ == self@.to_dynamic_object(),
+    {
+        // TODO: this might be unnecessarily slow
+        DynamicObject::from_kube(
+            deps_hack::k8s_openapi::serde_json::from_str(&deps_hack::k8s_openapi::serde_json::to_string(&self.inner).unwrap()).unwrap()
+        )
+    }
+
+    #[verifier(external_body)]
+    pub fn from_dynamic_object(obj: DynamicObject) -> (res: Result<ZookeeperCluster, ParseDynamicObjectError>)
+        ensures
+            res.is_Ok() == ZookeeperClusterView::from_dynamic_object(obj@).is_Ok(),
+            res.is_Ok() ==> res.get_Ok_0()@ == ZookeeperClusterView::from_dynamic_object(obj@).get_Ok_0(),
+    {
+        let parse_result = obj.into_kube().try_parse::<deps_hack::ZookeeperCluster>();
+        if parse_result.is_ok() {
+            let res = ZookeeperCluster { inner: parse_result.unwrap() };
+            Result::Ok(res)
+        } else {
+            Result::Err(ParseDynamicObjectError::ExecError)
+        }
+    }
+}
+
+impl ResourceWrapper<deps_hack::ZookeeperCluster> for ZookeeperCluster {
+    #[verifier(external)]
+    fn from_kube(inner: deps_hack::ZookeeperCluster) -> ZookeeperCluster {
+        ZookeeperCluster {
+            inner: inner
+        }
+    }
+
+    #[verifier(external)]
+    fn into_kube(self) -> deps_hack::ZookeeperCluster {
+        self.inner
+    }
+}
+
+#[verifier(external_body)]
+pub struct ZookeeperClusterSpec {
+    inner: deps_hack::ZookeeperClusterSpec,
+}
+
+impl ZookeeperClusterSpec {
+    pub spec fn view(&self) -> ZookeeperClusterSpecView;
+
+    #[verifier(external_body)]
+    pub fn replica(&self) -> (replica: i32)
+        ensures
+            replica as int == self@.replica,
+    {
+        self.inner.replica
+    }
+}
+
+impl ResourceWrapper<deps_hack::ZookeeperClusterSpec> for ZookeeperClusterSpec {
+    #[verifier(external)]
+    fn from_kube(inner: deps_hack::ZookeeperClusterSpec) -> ZookeeperClusterSpec {
+        ZookeeperClusterSpec {
+            inner: inner
+        }
+    }
+
+    #[verifier(external)]
+    fn into_kube(self) -> deps_hack::ZookeeperClusterSpec {
+        self.inner
+    }
+}
+
+}
