+++ src/controller_examples/zookeeper_controller/spec/zookeepercluster.rs
@@ -9,83 +9,9 @@ use vstd::prelude::*;
 
 verus! {
 
-#[verifier(external_body)]
-pub struct ZookeeperCluster {
-    inner: deps_hack::ZookeeperCluster
-}
-
 pub struct ZookeeperClusterView {
     pub metadata: ObjectMetaView,
     pub spec: ZookeeperClusterSpecView,
-}
-
-impl ZookeeperCluster {
-    pub spec fn view(&self) -> ZookeeperClusterView;
-
-    #[verifier(external_body)]
-    pub fn metadata(&self) -> (metadata: ObjectMeta)
-        ensures
-            metadata@ == self@.metadata,
-    {
-        ObjectMeta::from_kube(self.inner.metadata.clone())
-    }
-
-    #[verifier(external_body)]
-    pub fn replica(&self) -> (replica: i32)
-        ensures
-            replica as int == self@.spec.replica,
-    {
-        self.inner.spec.replica
-    }
-
-    #[verifier(external_body)]
-    pub fn api_resource() -> (res: ApiResource)
-        ensures
-            res@.kind == ZookeeperClusterView::kind(),
-    {
-        ApiResource::from_kube(deps_hack::kube::api::ApiResource::erase::<deps_hack::ZookeeperCluster>(&()))
-    }
-
-    // NOTE: This function assumes serde_json::to_string won't fail!
-    #[verifier(external_body)]
-    pub fn to_dynamic_object(self) -> (obj: DynamicObject)
-        ensures
-            obj@ == self@.to_dynamic_object(),
-    {
-        // TODO: this might be unnecessarily slow
-        DynamicObject::from_kube(
-            deps_hack::k8s_openapi::serde_json::from_str(&deps_hack::k8s_openapi::serde_json::to_string(&self.inner).unwrap()).unwrap()
-        )
-    }
-
-    #[verifier(external_body)]
-    pub fn from_dynamic_object(obj: DynamicObject) -> (res: Result<ZookeeperCluster, ParseDynamicObjectError>)
-        ensures
-            res.is_Ok() == ZookeeperClusterView::from_dynamic_object(obj@).is_Ok(),
-            res.is_Ok() ==> res.get_Ok_0()@ == ZookeeperClusterView::from_dynamic_object(obj@).get_Ok_0(),
-    {
-        let parse_result = obj.into_kube().try_parse::<deps_hack::ZookeeperCluster>();
-        if parse_result.is_ok() {
-            let res = ZookeeperCluster { inner: parse_result.unwrap() };
-            Result::Ok(res)
-        } else {
-            Result::Err(ParseDynamicObjectError::ExecError)
-        }
-    }
-}
-
-impl ResourceWrapper<deps_hack::ZookeeperCluster> for ZookeeperCluster {
-    #[verifier(external)]
-    fn from_kube(inner: deps_hack::ZookeeperCluster) -> ZookeeperCluster {
-        ZookeeperCluster {
-            inner: inner
-        }
-    }
-
-    #[verifier(external)]
-    fn into_kube(self) -> deps_hack::ZookeeperCluster {
-        self.inner
-    }
 }
 
 impl ZookeeperClusterView {
@@ -155,25 +81,8 @@ impl ResourceView for ZookeeperClusterView {
     proof fn from_dynamic_preserves_kind() {}
 }
 
-#[verifier(external_body)]
-pub struct ZookeeperClusterSpec {
-    inner: deps_hack::ZookeeperClusterSpec,
-}
-
 pub struct ZookeeperClusterSpecView {
     pub replica: int,
-}
-
-impl ZookeeperClusterSpec {
-    pub spec fn view(&self) -> ZookeeperClusterSpecView;
-
-    #[verifier(external_body)]
-    pub fn replica(&self) -> (replica: i32)
-        ensures
-            replica as int == self@.replica,
-    {
-        self.inner.replica
-    }
 }
 
 impl ZookeeperClusterSpecView {}
