+++ src/zookeeper_controller.rs
@@ -16,7 +16,7 @@ use builtin::*;
 use builtin_macros::*;
 
 use crate::zookeeper_controller::exec::reconciler::{ZookeeperReconcileState, ZookeeperReconciler};
-use crate::zookeeper_controller::spec::zookeepercluster::ZookeeperCluster;
+use crate::zookeeper_controller::exec::zookeepercluster::ZookeeperCluster;
 use deps_hack::anyhow::Result;
 use deps_hack::kube::CustomResourceExt;
 use deps_hack::serde_yaml;
